fix file upload

+++ src/libs/file_field/fields.py
@@ -15,10 +15,10 @@ except ImportError:
 
 class FieldFileMixin:
     def save(self, name, content, save=True):
-        old_name = os.path.normpath(self.name)
+        old_name = self.name and os.path.normpath(self.name)
         super().save(name, content, save=save)
         new_name = os.path.normpath(self.name)
-        if old_name != new_name:
+        if old_name and old_name != new_name and os.path.isfile(old_name):
             self.storage.delete(old_name)
     save.alters_data = True
